logic divided to repository and service
