+++ app/repositories/__init__.py
@@ -1,0 +1,5 @@
+# app/repositories/__init__.py
+
+from .trading_repository import TradingRepository
+
+__all__ = ["TradingRepository"]
