+++ app/services/__init__.py
@@ -1,0 +1,5 @@
+# app/services/__init__.py
+
+from .trading_service import TradingService
+
+__all__ = ["TradingService"]
